fix: nix.gc for darwin
On Darwin, the old implementation added the entire "options" attribute as a
single entry in the ProgramArguments.  This yields e.g.:

    nix.gc = {
      automatic = true;
      frequency = "daily";
      options = "--delete-older-than 30d";
    };

=>

	<key>ProgramArguments</key>
	<array>
		<string>/nix/store/ckm9hpnd4fym65mkypbk7rmzp7m6ad0v-nix-2.24.10/bin/nix-collect-garbage</string>
		<string>--delete-older-than 30d</string>
	</array>

while what you really want is:

	<key>ProgramArguments</key>
	<array>
		<string>/nix/store/ckm9hpnd4fym65mkypbk7rmzp7m6ad0v-nix-2.24.10/bin/nix-collect-garbage</string>
		<string>--delete-older-than</string>
		<string>30d</string>
	</array>

Or just wrap it in a script the same way as is done for linux.  In fact that
script is completely compatible with darwin, so we can just reuse the existing one.